dnsproxy: Be more strict with incoming DNS requests
Be more strict with incoming DNS requests. Verify that there is
only one entry in the question section and none in the answer or
name server sections.

Ensure that the question section has a proper IN class and
compute the remaing length and the pointer position with respect
to the end of the question section. If there is an EDNS0 extension,
log its length.