service: Do no leak stats file descriptors
ConnMan leaks the stats file descriptor. The disconnect path is
slightly more complex then it one would suspect. So when the service
goes from ONLINE to IDLE because the Ethernet cable was unplugged, the
service will retry to connect. This ends up opening up another FD for
the stats file. Eventually the service will be removed but we bail out
a early in the __connman_service_disconnect() function and never call
__connann_stats_servicde_unregister. Let's move the unregister up so
that we always call it.