Fix WiFi enable not working.
On system startup there is an error message from connman "Failed to open RFKILL control device". This turns out to be caused by the RFKILL menu to be loaded when connman starts, that's too late. We added a modprobe command as a prestart command to ensure this dependency is fixed.
Another solution would have been to include the RFKILL module into the Linux kernel, but then the dependency is implicit and not documented.

EM-3184