Handle rfkill module dependency with systemd module load service
Wifi not working was caused by a race condition on the rfkill device
which could not be opened because it is loaded to late. The fix
was implementate to use systemd's ExexPreStart functionality to
modprobe the module first.
A better more maintainable approach is using the systemd's module load
facility.

'systemd-modules-load.service is an early boot service that loads
kernel modules based on static configuration.'

Solves EMP-638, contributes to EM-3184

Signed-off-by: Raymond Siudak <r.siudak@ultimaker.com>