EM-3619 Disable the background scanning by default
Background scanning is intended for roaming, i.e. a user moving his device to other locations
will then be able to automatically switch to an access point with a stronger signal.
Useless for our printer and a potential source of errors.

This feature introduced problems in the WiFi setup. The WiFi setup restarts Connman
before connecting to the new selected WiFi network. At Connman startup the scans repeat with short intervals
(which become longer over time). A scan also takes long (about 5 seconds) and then doesn't handle DBUS calls. The
problem is that when background scanning is enabled we don't have control about when these scans will
happen and it has happened this did occur at the same time we wanted to connect to the new WiFi network,
resulting in a DBus timeout. Lacking DBus error recovery resulted in the WiFi setup failing.

Disabling the background scanning should result in just a single scan when requested.

Normally we would disable this feature in the configuration file, but the new feature for the user
configuring IP-addresses by overwriting that configuration file complicates things. Changing the default
value seems like the easiest solution