EM-3619 Disable the online check
Connman by default does a (periodic) check for the internet connection being active, here we disable that feature.

Users have asked questions for why the printer is contacting a weird domain address. It would make more sense to
check for ultimaker.com being accessible. Since we don't use this feature it feels better to disable it.

Normally we would disable this feature in the configuration file, but the new feature for the user
configuring IP-addresses by overwriting that configuration file complicates things. Changing the default
value seems like the easiest solution